Refactor admin pages: rename components and update imports for consistency
